+++ dart/tool/dart/generate_dart.dart
@@ -1324,7 +1324,14 @@ class Type extends Member {
 
     fields.forEach((TypeField field) {
       if (field.type.isSimple || field.type.isEnum) {
-        gen.write("${field.generatableName} = json['${field.name}']");
+        // Special case `AllocationProfile`.
+        if (name == 'AllocationProfile' && field.type.name == 'int') {
+          gen.write(
+              "${field.generatableName} = json['${field.name}'] is String ? "
+                  "int.parse(json['${field.name}']) : json['${field.name}']");
+        } else {
+          gen.write("${field.generatableName} = json['${field.name}']");
+        }
         if (field.defaultValue != null) {
           gen.write(' ?? ${field.defaultValue}');
         }
@@ -1388,8 +1395,15 @@ class Type extends Member {
             gen.writeln("${field.generatableName} = "
                 "new List<${fieldType.listTypeArg}>.from($ref);");
           } else {
-            gen.writeln("${field.generatableName} = "
-                "new List<${fieldType.listTypeArg}>.from(createServiceObject($ref));");
+            // Special case `InstanceSet`. Pre 3.20, instances were sent in a
+            // field named 'samples' instead of 'instances'.
+            if (name == 'InstanceSet') {
+              gen.writeln("${field.generatableName} = "
+                  "new List<${fieldType.listTypeArg}>.from(createServiceObject($ref ?? json['samples']));");
+            } else {
+              gen.writeln("${field.generatableName} = "
+                  "new List<${fieldType.listTypeArg}>.from(createServiceObject($ref));");
+            }
           }
         }
       } else {