+++ dart/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.20.0+1
 - handle null isolate ids in `callServiceExtension`
+- add backwards compatibility for `InstanceSet` and `AllocationProfile`
 
 ## 3.20.0
 - rev to 3.20.0; expose public methods added in 3.17 - 3.20 VM Service Protocol versions